# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateLibraryRequest:
    properties:
      name:
        type: string
      path:
        type: string
      type:
        type: string
    type: object
...
  handlers.LibraryResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      path:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  handlers.LibrarySourceResponse:
    properties:
      api_key:
        type: string
      cookies:
        type: string
      enabled:
        type: boolean
      magazine_blocklist:
        items:
          type: string
        type: array
      misc_blocklist:
        items:
          type: string
        type: array
      source:
        type: string
    type: object
...
  handlers.UpdateLibraryRequest:
    properties:
      name:
        type: string
    type: object
  handlers.UpdateLibrarySourceRequest:
    properties:
      api_key:
        type: string
      cookies:
        type: string
      enabled:
        type: boolean
      magazine_blocklist:
        items:
          type: string
        type: array
      misc_blocklist:
        items:
          type: string
        type: array
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Self-hosted doujinshi library manager API
  title: Shoka API
  version: "1.0"
paths:
...
  /api/admin/libraries:
    post:
      consumes:
      - application/json
      parameters:
      - description: Library to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateLibraryRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.LibraryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create a library
      tags:
      - admin
  /api/admin/libraries/{id}:
    delete:
      description: Cascade-deletes every archive and cached data belonging to this
        library. Does not touch files on disk.
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete a library
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: The library's path and type are immutable after creation.
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLibraryRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LibraryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Rename a library
      tags:
      - admin
  /api/admin/libraries/{id}/scan:
    post:
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Trigger a scan of a single library
      tags:
      - admin
  /api/admin/libraries/{id}/sources:
    get:
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.LibrarySourceResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get a library's per-library metadata source settings
      tags:
      - admin
  /api/admin/libraries/{id}/sources/{source}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      - description: Source name (e.g. nhentai, e-hentai, comicinfo, filename)
        in: path
        name: source
        required: true
        type: string
      - description: Source settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLibrarySourceRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LibrarySourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Configure a metadata source for a single library
      tags:
      - admin
...
  /api/libraries:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.LibraryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List libraries
      tags:
      - libraries
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"